requirements.txt
redis
numpy
pygame
pyyaml